feat: make rm.js
